feat: Create NewLessonButton component and update the types Lesson and Course in lib/types
