+++ app/(dashboard)/(routes)/courses/loading.tsx
@@ -5,18 +5,16 @@ const Loading = async () => {
 
   return (
     <main className="h-full w-full p-4">
-      <section className="w-full p-2 items-center grid sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
+      <section className="grid w-full items-center p-2 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4">
         {courses.map((course) => (
           <Skeleton
             key={course.id}
-            className="min-w-[10rem] max-w-xl h-72 p-2 flex flex-col rounded-md cursor-pointer border-2  m-2"
+            className="m-2 flex h-72 min-w-[10rem] max-w-xl cursor-pointer flex-col rounded-md border-2  p-2"
           >
-            <Skeleton className=" w-full h-64 rounded-md bg-gray-200">
-              .
-            </Skeleton>
+            <Skeleton className=" h-64 w-full rounded-md bg-gray-200"></Skeleton>
             <Skeleton className="  rounded-b-md">
-              <Skeleton className="my-4 font-medium text-xl"></Skeleton>
-              <Skeleton className="text-slate-500 inline-flex items-center ">
+              <Skeleton className="my-4 text-xl font-medium"></Skeleton>
+              <Skeleton className="inline-flex items-center text-slate-500 ">
                 <Skeleton className="h-4 w-4 " />
               </Skeleton>
             </Skeleton>
