+++ app/(dashboard)/(routes)/teacher/course/_components/publishCourseButton.tsx
@@ -3,28 +3,7 @@ import React, { useState } from "react";
 import { Switch } from "@/components/ui/switch";
 import axios from "axios";
 import toast from "react-hot-toast";
-
-type CourseData = {
-  id: string;
-  title: string;
-  description: string | null;
-  createdAt: Date;
-  updateAt: Date;
-  imageUrl: string | null;
-  isPublished: boolean;
-  lessons: Lesson[];
-} | null;
-
-type Lesson = {
-  id: string;
-  title: string;
-  description: string;
-  createdAt: Date;
-  updateAt: Date;
-  videoUrl: string;
-  courseId: string;
-  isPublished: boolean;
-};
+import { CourseData } from "@/lib/types";
 
 const PublishCourseButton = ({ courseData }: { courseData: CourseData }) => {
   const [toggle, setToggle] = useState<boolean | undefined>(
