+++ app/(dashboard)/(routes)/teacher/page.tsx
@@ -18,7 +18,7 @@ const TeacherPage = async () => {
         <TableCaption>A list of your Courses.</TableCaption>
         <TableHeader>
           <TableRow>
-            <TableHead className="text-left">Title</TableHead>
+            <TableHead className="text-left">Course</TableHead>
             <TableHead className="text-center">Status</TableHead>
             <TableHead className="text-center">Lessons</TableHead>
           </TableRow>
